+++ outputs.tf
@@ -1,0 +1,7 @@
+output "arn" {
+  value = "${local.arn}"
+}
+
+output "function_name" {
+  value = "${module.lambda-label.function_name}"
+}